+++ libs/math/Plane3.cpp
@@ -3,7 +3,7 @@
 #include "AABB.h"
 #include "Matrix4.h"
 
-float Plane3::distanceToOrientedExtents(const Vector3& extents, const Matrix4& orientation) const
+double Plane3::distanceToOrientedExtents(const Vector3& extents, const Matrix4& orientation) const
 {
 	return fabs(extents[0] * normal().dot(orientation.x().getVector3())) + 
 		   fabs(extents[1] * normal().dot(orientation.y().getVector3())) + 
@@ -12,7 +12,7 @@ float Plane3::distanceToOrientedExtents(const Vector3& extents, const Matrix4& o
 
 bool Plane3::containsAABB(const AABB& aabb, const Matrix4& orientation) const
 {
-	float dot = distanceToPointAABB(aabb.origin);
+	double dot = distanceToPointAABB(aabb.origin);
 
 	return !(dot > 0 || -dot < distanceToOrientedExtents(aabb.extents, orientation));
 }
